+++ src/MetaEstimator.php
@@ -4,5 +4,10 @@ namespace Rubix\ML;
 
 interface MetaEstimator extends Estimator
 {
-    //
+    /**
+     * Return the base estimator instance.
+     *
+     * @return \Rubix\ML\Estimator
+     */
+    public function estimator() : Estimator;
 }