+++ src/Other/Traits/LoggerAware.php
@@ -2,6 +2,8 @@
 
 namespace Rubix\ML\Other\Traits;
 
+use Rubix\ML\Verbose;
+use Rubix\ML\MetaEstimator;
 use Psr\Log\LoggerInterface;
 
 trait LoggerAware
@@ -21,6 +23,14 @@ trait LoggerAware
      */
     public function setLogger(LoggerInterface $logger) : void
     {
+        if ($this instanceof MetaEstimator) {
+            $estimator = $this->estimator();
+            
+            if ($estimator instanceof Verbose) {
+                $estimator->setLogger($logger);
+            }
+        }
+
         $this->logger = $logger;
     }
 }